requirements_transcriber.txt
groq
python-dotenv
PyMuPDF
Pillow
pytesseract
langchain
pdfplumber